Add an AccountOperations interface
A high level description of the operations we wish to perform.
We start with creating an account.
We think we will need a name for the owner and a password to access it.
This is within a local schema, so the "account number" (IBAN, BIC etc.)
will be generated automatically.

A specific "account number", called 'accountCode', can be given with the
second method of this interface.

All these parameters are strings for future RegEx patterns.

(NOTE: we implicitly return the DTO when creating an account)