Add Account entity
While the account code is a String field of this class, the rest of the
Strings are wrapped in different classes (Credentials for the password,
OwnerInfo for the account owner's name).

This is so that these objects may be more detailed in the future.
Also, they should probably be stored in different tables
than the account info.

We also see that the balance usually requires a unit of measure, here
the currency.
Currency is an enum, and has default value Euros.