Add other methods to AccountOperations
Specifically, we will be getting Accounts by Account.id and Account.code
(This is because we have set the code to be stored with a
Unique constraint)

Then we will deposit/withdraw. These have a separate String functioning as
a payload to be checked. (The scenario envisioned is that one can not
withdraw/deposit money with a simple curl/Postman request - these need to
be done at a specific authorized point: ATM, proprietary app,
brick & mortar bank etc. While beyond the scope of this MVP, they
can be used in the future to confirm this authorization)

The InsufficientFundsException and UnauthorizedOperationException
come to mind from the transfer/withdrawal/deposit points above.

Transferring amounts between accounts, however, is something that
can be done within the API itself, via a programmatic request and should
not need this verification (as long as transfers are traceable).

Displaying the balance returns a double, not a long.
The long is stored as a fixed point algebra value,
(the precise value in the centimes of the currency)
while some rounding logic (again, something to be
extended in the future) is needed to convert it to this
floating point algebra Double.
(simply dividing by 100 and truncating should be OK for
now)